refactor: move transport settings and build_async_openai to clients.client
Timeouts, limits, and the retry policy are client behavior, not endpoint
schema; configs.client keeps only the config classes and API-key resolution.
